# src/main/java/sgcp/transactions/manager/fimpe/operation/dto/OperationTypeDTO.java
package sgcp.transactions.manager.fimpe.operation.dto;

import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
public class OperationTypeDTO {

    private Long id;
    private String nameOperationType;


}