# src/main/java/sgcp/transactions/manager/fimpe/operation/dto/PersonTypeDTO.java
package sgcp.transactions.manager.fimpe.operation.dto;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
public class PersonTypeDTO {

    private Long id;
    private String namePersonType;

}